认证方式完善

+++ app/apps/api/users/users.py
@@ -1,11 +1,12 @@
 import time
 from typing import Union, Optional
 from fastapi import APIRouter, HTTPException, UploadFile, File, Form, Request, Depends, Security
+from fastapi.security import HTTPAuthorizationCredentials
 from redis import Redis
 from tortoise.contrib.pydantic import pydantic_model_creator
 from tortoise.expressions import Q
 
-from apps.dependencies.auth import get_current_user, bearer_scheme
+from apps.dependencies.auth import get_current_user, bearer_scheme, get_token_str
 from apps.form.users.form import UserCreate, TokenResponse
 from apps.models.user import User
 from apps.utils import response
@@ -20,13 +21,8 @@ router = APIRouter(prefix="/users", tags=["用户管理"])
 User_Pydantic = pydantic_model_creator(User, name="User", exclude=("hashed_password",))
 
 
-
-@router.get("/list", summary="用户列表", description="获取用户列表", dependencies=[Security(bearer_scheme)])
-async def user_list(
-        username: Optional[str] = None,
-        page: int = 1,
-        page_size: int = 10
-):
+@router.get("/list", summary="用户列表", description="获取用户列表", dependencies=[Depends(get_current_user)])
+async def user_list(username: Optional[str] = None, page: int = 1, page_size: int = 10):
     conditions = []
     if username:
         conditions.append(Q(username__icontains=username))
@@ -39,7 +35,8 @@ async def user_list(
     return response(data=data, total=total, total_page=total_page, message="获取用户列表成功！")
 
 
-@router.post("/register/", response_model=User_Pydantic, summary="注册用户", description="创建用户接口")
+@router.post("/register/", response_model=User_Pydantic, summary="注册用户", description="创建用户接口",
+             dependencies=[Depends(get_current_user)])
 async def create_user(user: UserCreate):
     # 判断用户名或邮箱是否已经被注册
     if await User.filter(Q(username=user.username) | Q(email=user.email)).exists():
@@ -55,7 +52,8 @@ async def create_user(user: UserCreate):
     return response(data=data.model_dump(), message="注册成功！")
 
 
-@router.put("/update/{user_id}", response_model=User_Pydantic, summary="更新用户", description="更新用户信息")
+@router.put("/update/{user_id}", response_model=User_Pydantic, summary="更新用户", description="更新用户信息",
+            dependencies=[Depends(get_current_user)])
 async def update_user(user_id: int, user: UserCreate):
     if await User.filter(Q(username=user.username) | Q(email=user.email)).exclude(id=user_id).exists():
         return response(code=0, message="用户名或邮箱已经被注册！")
@@ -74,7 +72,8 @@ async def update_user(user_id: int, user: UserCreate):
     return response(data=data.model_dump(), message="更新成功！")
 
 
-@router.get("/detail/{user_id}", response_model=User_Pydantic, summary="用户详情", description="获取用户详情")
+@router.get("/detail/{user_id}", response_model=User_Pydantic, summary="用户详情", description="获取用户详情",
+            dependencies=[Depends(get_current_user)])
 async def read_user(user_id: int):
     user = await User.get_or_none(id=user_id)
     if user is None:
@@ -83,18 +82,17 @@ async def read_user(user_id: int):
     return response(data=data.model_dump())
 
 
-@router.delete("/delete/{user_id}", response_model=dict, summary="删除用户", description="删除用户")
-async def delete_user(user_id: int):
-    deleted_count = await User.filter(id=user_id).delete()
-    return response(data={"deleted_count": deleted_count})
+@router.delete("/delete/{user_id}", response_model=dict, summary="删除用户", description="删除用户",
+               dependencies=[Depends(get_current_user)])
+async def delete_user(user_id: int, user: User = Depends(get_current_user)):
+    if user.id == user_id:
+        return response(code=0, message="不允许删除自身账号！")
+    user_id = await User.filter(id=user_id).delete()
+    return response(data={"id": user_id})
 
 
 @router.post("/login", summary="登录接口", response_model=TokenResponse, description="登录接口")
-async def login(
-        username: str = Form(...),
-        password: str = Form(...),
-        redis_client: Redis = Depends(get_redis_client)  # 自动注入
-):
+async def login(username: str = Form(...), password: str = Form(...), redis_client: Redis = Depends(get_redis_client)):
     # 这里可以添加登录逻辑
     print(username, password)
     decrypt_pwd = decrypt(AES_KEY, password)
@@ -113,10 +111,10 @@ async def login(
     return response(data=resp, message="登录成功！")
 
 
-@router.get("/logout", summary="注销接口", description="注销接口")
+@router.get("/logout", summary="注销接口", description="注销接口", dependencies=[Depends(get_current_user)])
 async def logout(
         redis_client: Redis = Depends(get_redis_client),
-        token: str = Security(bearer_scheme)
+        token: str = Depends(get_token_str),
 ):
     if not token:
         return response(code=0, message="未登录")
